+++ commodore/cluster.py
@@ -131,8 +131,6 @@ def render_target(
     classes = [f"params.{inv.bootstrap_target}"]
     parameters: Dict[str, Union[Dict, str]] = {
         "_instance": target,
-        "component_versions": {},
-        "components": "${component_versions}",
     }
 
     for c in components:
@@ -206,6 +204,14 @@ def render_params(inv: Inventory, cluster: Cluster):
             "customer": {
                 "name": cluster.tenant,
             },
+            # Merge component_versions into components in params.cluster for
+            # backwards-compatibility.
+            # We do this here instead of the target to ensure values in
+            # `components` have precedence over values in
+            # `component_versions`.
+            # TODO Remove once the deprecated `component_versions` field is removed
+            "component_versions": {},
+            "components": "${component_versions}",
         },
     }